requirements.txt
GitPython
requests
pyperclip
python-dotenv
